# openapi.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: A simple CRUD API for managing a library book collection
  version: 1.0.0
  contact:
    name: Library API Support
  license:
    name: ISC

servers:
  - url: http://localhost:3001
    description: Local development server

tags:
  - name: Books
    description: Operations related to book management

paths:
  /api/books:
    get:
      tags:
        - Books
      summary: Get all books
      description: Retrieve a list of all books in the library
      operationId: getAllBooks
      responses:
        '200':
          description: Successfully retrieved list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
              example:
                - id: "550e8400-e29b-41d4-a716-446655440001"
                  title: "To Kill a Mockingbird"
                  author: "Harper Lee"
                  year: 1960
                  isbn: "978-0061120084"
                - id: "550e8400-e29b-41d4-a716-446655440002"
                  title: "1984"
                  author: "George Orwell"
                  year: 1949
                  isbn: "978-0451524935"

    post:
      tags:
        - Books
      summary: Create a new book
      description: Add a new book to the library collection
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
            example:
              title: "The Great Gatsby"
              author: "F. Scott Fitzgerald"
              year: 1925
              isbn: "978-0743273565"
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
              example:
                id: "550e8400-e29b-41d4-a716-446655440004"
                title: "The Great Gatsby"
                author: "F. Scott Fitzgerald"
                year: 1925
                isbn: "978-0743273565"
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Title and author are required"

  /api/books/{id}:
    get:
      tags:
        - Books
      summary: Get a book by ID
      description: Retrieve details of a specific book
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the book to retrieve
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Successfully retrieved book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
              example:
                id: "550e8400-e29b-41d4-a716-446655440001"
                title: "To Kill a Mockingbird"
                author: "Harper Lee"
                year: 1960
                isbn: "978-0061120084"
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Book not found"

    put:
      tags:
        - Books
      summary: Update a book
      description: Update an existing book's information
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the book to update
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
            example:
              title: "To Kill a Mockingbird (Updated)"
              author: "Harper Lee"
              year: 1960
              isbn: "978-0061120084"
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
              example:
                id: "550e8400-e29b-41d4-a716-446655440001"
                title: "To Kill a Mockingbird (Updated)"
                author: "Harper Lee"
                year: 1960
                isbn: "978-0061120084"
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Title and author are required"
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Book not found"

    delete:
      tags:
        - Books
      summary: Delete a book
      description: Remove a book from the library collection
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the book to delete
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Book deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
              example:
                message: "Book deleted successfully"
                book:
                  id: "550e8400-e29b-41d4-a716-446655440001"
                  title: "To Kill a Mockingbird"
                  author: "Harper Lee"
                  year: 1960
                  isbn: "978-0061120084"
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Book not found"

components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - author
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the book (UUID)
          example: "550e8400-e29b-41d4-a716-446655440001"
        title:
          type: string
          description: Title of the book
          example: "To Kill a Mockingbird"
        author:
          type: string
          description: Author of the book
          example: "Harper Lee"
        year:
          type: integer
          nullable: true
          description: Publication year
          example: 1960
        isbn:
          type: string
          description: ISBN number
          example: "978-0061120084"

    BookInput:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          description: Title of the book
          example: "To Kill a Mockingbird"
        author:
          type: string
          description: Author of the book
          example: "Harper Lee"
        year:
          type: integer
          nullable: true
          description: Publication year
          example: 1960
        isbn:
          type: string
          description: ISBN number
          example: "978-0061120084"

    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Book not found"

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: "Book deleted successfully"
        book:
          $ref: '#/components/schemas/Book'